refactor(gtf): move test cases to _test file
